pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "extended-mypy-django-plugin"
dynamic = ["version"]
description = "Trying to make mypy understand .objects on abstract django models"
readme = "README.rst"
license = { text = "MIT" }
authors = [
    { name = "Stephen Moore", email = "stephen@delfick.com" },
]

[project.optional-dependencies]
tests = [
    "pytest==8.1.1",
]
stubs-latest = [
    "mypy==1.10.0",
    "django-stubs==5.0.0",
]
stubs-older = [
    "mypy==1.4.0",
    "django-stubs==4.2.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "extended_mypy_django_plugin/version.py"

[tool.hatch.build.targets.wheel]
include = [
    "/extended_mypy_django_plugin",
]

[tool.ruff]
target-version = "py310"
line-length = 99
extend-exclude = [
    ".DS_Store",
    ".pytest_cache",
    "tools/.python",
    "tools/deps",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "B008",  # mutable objects as function argument defaults
    "B015",  # flake8-bugbear (useless-comparison)
    "I",  # isort
    "T20",  # flake8-print
    "UP",  # pyupgrade
    "RUF",
    "F",
]

[tool.ruff.lint.per-file-ignores]
"example/djangoexample/views.py" = ["F821"]
"extended_mypy_django_plugin/plugin/_debug.py" = ["T201"]
